Optimizations for Infinispan 7

+++ guide/chapter12-cdi/src/main/java/com/packtub/infinispan/chapter12/manager/OpenWeatherCacheInfo.java
@@ -40,7 +40,7 @@ public class OpenWeatherCacheInfo {
 		return cache.getCacheConfiguration().eviction().strategy();
 	}
 
-	public int getEvictionMaxEntries() {
+	public long getEvictionMaxEntries() {
 		return cache.getCacheConfiguration().eviction().maxEntries();
 	}
